+++ sensor/src/main/java/medvedev/ilya/monitor/sensor/model/SensorValue.java
@@ -2,9 +2,9 @@ package medvedev.ilya.monitor.sensor.model;
 
 public class SensorValue {
     private final String name;
-    private final double value;
+    private final float value;
 
-    public SensorValue(final String name, final double value) {
+    public SensorValue(final String name, final float value) {
         this.name = name;
         this.value = value;
     }
@@ -13,7 +13,7 @@ public class SensorValue {
         return name;
     }
 
-    public double getValue() {
+    public float getValue() {
         return value;
     }
 }
